refactor: extract chat session selection drag state
