+++ dashboard/src/utils/sessionManagement.mjs
@@ -1,6 +1,4 @@
 export const DRAG_MIME_SESSION_IDS = "application/x-astrbot-session-ids";
-export const DRAG_MIME_SOURCE_PROJECT_ID =
-  "application/x-astrbot-source-project-id";
 
 export function toggleSessionSelection(selectedSessionIds, sessionId) {
   return selectedSessionIds.includes(sessionId)
@@ -19,10 +17,6 @@ export function shouldSuppressClickAfterLongPress(suppressNextClick) {
     suppress: suppressNextClick,
     nextSuppressState: false,
   };
-}
-
-export function toggleExpandedProject(currentProjectId, projectId) {
-  return currentProjectId === projectId ? null : projectId;
 }
 
 export function toggleExpandedProjectIds(currentProjectIds, projectId) {
